Move visualization stuff into a new directory called visualization
