+++ applications/compare_cumulative_psf.py
@@ -86,10 +86,11 @@ import numpy as np
 import yaml
 
 import simtools.util.general as gen
-from simtools import io_handler, visualize
+from simtools import io_handler
 from simtools.configuration import configurator
 from simtools.model.telescope_model import TelescopeModel
 from simtools.ray_tracing import RayTracing
+from simtools.visualization import visualize
 
 
 def load_data(datafile):
