+++ simtools/camera_efficiency.py
@@ -8,10 +8,11 @@ import numpy as np
 from astropy.table import Table
 
 import simtools.util.general as gen
-from simtools import io_handler, visualize
+from simtools import io_handler
 from simtools.model.telescope_model import TelescopeModel
 from simtools.simtel.simtel_runner_camera_efficiency import SimtelRunnerCameraEfficiency
 from simtools.util import names
+from simtools.visualization import visualize
 
 __all__ = ["CameraEfficiency"]
