+++ simtools/visualization/legend_handlers.py
