+++ simtools/visualization/visualize.py
